+++ webstepper/CHANGELOG.md
@@ -17,6 +17,8 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### 🔧 Internal changes
 
+- Removed the unused `automation` parameter when calling the draw function
+
 ## [0.9.1] - 2026-06-11
 
 ### ✨ Enhancements